Check valid emails with 'parse'
This allows us to be confident that we're getting the correct results
from the parsing as well as treating the email as valid. Otherwise we
might risk parse succeeding but with the wrong results.

As parse doesn't succeed with invalid emails addresses we leave those as
being tested with 'isValid'.